CpuContext.addWatcher was Cpu
- TODO Move Cpu.addBreakpoint
