+++ src/main/java/walkingkooka/emulator/c64/Cpu.java
@@ -25,8 +25,6 @@ public interface Cpu {
 
     Runnable addBreakpoint(final short address);
 
-    Runnable addWatcher(final CpuWatcher watcher);
-
     String disassemble(final CpuContext context);
 
     void step(final CpuContext context);
