+++ src/main/java/walkingkooka/emulator/c64/CpuTesting.java
@@ -28,15 +28,6 @@ public interface CpuTesting<C extends Cpu> extends TreePrintableTesting,
     ClassTesting<C> {
 
     @Test
-    default void testAddWatcherWithNullFails() {
-        assertThrows(
-            NullPointerException.class,
-            () -> this.createCpu()
-                .addWatcher(null)
-        );
-    }
-
-    @Test
     default void testDisassembleWithNullContextFails() {
         assertThrows(
             NullPointerException.class,
